backend/requirements.txt
# Core
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20
pydantic==2.10.4
pydantic-settings==2.7.1

# Google AI
google-genai==1.0.0

# Document Processing
PyMuPDF==1.25.3
python-pptx==1.0.2
python-docx==1.1.2

# Vector Database
chromadb==0.6.3

# Utilities
aiofiles==24.1.0
python-dotenv==1.0.1
posthog>=2.4.0,<6.0.0
